+++ tests/packagedcode/test_debian.py
@@ -34,7 +34,7 @@ class TestDebianPackageGetInstalledPackages(PackageTester):
 
     def test_can_scan_system_package_installed_debian_with_license_from_container_layer(self):
         test_dir = self.extract_test_tar('debian/debian-container-layer.tar.xz')
-        expected_file = self.get_test_loc('debian/debian-container-layer.tar.xz.scan-expected.json', must_exist=False)
+        expected_file = self.get_test_loc('debian/debian-container-layer.tar.xz.scan-expected.json')
         result_file = self.get_temp_file('results.json')
         run_scan_click(['--system-package', test_dir, '--json-pp', result_file])
         check_json_scan(expected_file, result_file, regen=REGEN_TEST_FIXTURES)
@@ -42,7 +42,7 @@ class TestDebianPackageGetInstalledPackages(PackageTester):
     def test_can_get_installed_system_packages_with_license_from_debian_container_layer(self):
         from packagedcode.plugin_package import get_installed_packages
         test_dir = self.extract_test_tar('debian/debian-container-layer.tar.xz')
-        expected_file = self.get_test_loc('debian/debian-container-layer.tar.xz.get-installed-expected.json', must_exist=False)
+        expected_file = self.get_test_loc('debian/debian-container-layer.tar.xz.get-installed-expected.json')
         results = list(get_installed_packages(test_dir))
         self.check_packages_data(results, expected_file, remove_uuid=True, regen=REGEN_TEST_FIXTURES)
 
@@ -70,6 +70,15 @@ class TestDebian(PackageTester):
         packages = list(debian.DebianInstalledStatusDatabaseHandler.parse(test_file))
         self.check_packages_data(packages, expected_loc, regen=REGEN_TEST_FIXTURES)
 
+    def test_parse_status_file_basic_package_only(self):
+        test_file = self.get_test_loc('debian/basic/status')
+        expected_loc = self.get_test_loc('debian/basic/status-package-only.expected')
+        # specify ubuntu distro as this was the source of the test `status` file
+        packages = list(debian.DebianInstalledStatusDatabaseHandler.parse(
+            location=test_file, package_only=True,
+        ))
+        self.check_packages_data(packages, expected_loc, regen=REGEN_TEST_FIXTURES, package_only=True)
+
     def test_parse_status_file_with_source_packages(self):
         test_file = self.get_test_loc('debian/status-with-source/status')
         expected_loc = self.get_test_loc('debian/status-with-source/status.expected')
@@ -85,7 +94,7 @@ class TestDebian(PackageTester):
     @skipIf(on_windows, 'File names cannot contain colons on Windows')
     def test_scan_system_package_end_to_end_installed_debian(self):
         test_dir = self.extract_test_tar('debian/end-to-end.tgz')
-        expected_file = self.get_test_loc('debian/end-to-end.tgz.expected.json', must_exist=False)
+        expected_file = self.get_test_loc('debian/end-to-end.tgz.expected.json')
         result_file = self.get_temp_file('results.json')
         run_scan_click(['--system-package', test_dir, '--json-pp', result_file])
         check_json_scan(expected_file, result_file, regen=REGEN_TEST_FIXTURES)
@@ -137,7 +146,7 @@ class TestDebianGetListOfInstalledFiles(PackageTester):
             package_type=debian.DebianInstalledMd5sumFilelistHandler.default_package_type,
         )
 
-        expected_loc = self.get_test_loc('debian/libatk-adaptor-amd64.md5sums.expected.json', must_exist=False)
+        expected_loc = self.get_test_loc('debian/libatk-adaptor-amd64.md5sums.expected.json')
         self.check_packages_data(results, expected_loc, must_exist=False, regen=REGEN_TEST_FIXTURES)
 
     def test_parse_debian_files_list_no_arch(self):
@@ -148,7 +157,7 @@ class TestDebianGetListOfInstalledFiles(PackageTester):
             package_type=debian.DebianInstalledMd5sumFilelistHandler.default_package_type,
         )
 
-        expected_loc = self.get_test_loc('debian/files-md5sums/mokutil.md5sums.expected.json', must_exist=False)
+        expected_loc = self.get_test_loc('debian/files-md5sums/mokutil.md5sums.expected.json')
         self.check_packages_data(results, expected_loc, must_exist=False, regen=REGEN_TEST_FIXTURES)
 
     @skipIf(on_windows, 'File names cannot contain colons on Windows')
@@ -167,7 +176,7 @@ class TestDebianGetListOfInstalledFiles(PackageTester):
             package_type=debian.DebianInstalledMd5sumFilelistHandler.default_package_type,
         )
 
-        expected_loc = self.get_test_loc('debian/files-md5sums/mokutil-amd64.md5sums.expected.json', must_exist=False)
+        expected_loc = self.get_test_loc('debian/files-md5sums/mokutil-amd64.md5sums.expected.json')
         self.check_packages_data(results, expected_loc, must_exist=False, regen=REGEN_TEST_FIXTURES)
 
     def test_build_package_data_from_package_filename_deb_does_not_crash_on_version(self):